Move nonpredictive recommender subclasses to subpackage
