+++ baybe/recommenders/__init__.py
@@ -4,12 +4,12 @@ from baybe.recommenders.bayesian import (
     NaiveHybridRecommender,
     SequentialGreedyRecommender,
 )
-from baybe.recommenders.clustering import (
+from baybe.recommenders.nonpredictive.clustering import (
     GaussianMixtureClusteringRecommender,
     KMeansClusteringRecommender,
     PAMClusteringRecommender,
 )
-from baybe.recommenders.sampling import FPSRecommender, RandomRecommender
+from baybe.recommenders.nonpredictive.sampling import FPSRecommender, RandomRecommender
 
 __all__ = [
     "FPSRecommender",
