+++ baybe/recommenders/nonpredictive/__init__.py
@@ -1,1 +1,16 @@
 """Nonpredictive recommenders."""
+
+from baybe.recommenders.nonpredictive.clustering import (
+    GaussianMixtureClusteringRecommender,
+    KMeansClusteringRecommender,
+    PAMClusteringRecommender,
+)
+from baybe.recommenders.nonpredictive.sampling import FPSRecommender, RandomRecommender
+
+__all__ = [
+    "FPSRecommender",
+    "GaussianMixtureClusteringRecommender",
+    "KMeansClusteringRecommender",
+    "PAMClusteringRecommender",
+    "RandomRecommender",
+]
